Moved `registerOperationResources` by Router into base class and removed `AbstractAnnotationConfigurationLoader` as redundant.
